Add doc input `Month::class`.
