+++ docs/form/input/Time.md
@@ -128,7 +128,6 @@ The following methods are available for setting attributes:
 | `lang()`           | Set the `lang` attribute.                                                                       |
 | `name()`           | Set the `name` attribute.                                                                       |
 | `readOnly()`       | Set the `readonly` attribute.                                                                   |
-| `step()`           | Set the `step` attribute.                                                                       |
 | `style()`          | Set the `style` attribute.                                                                      |
 | `tabIndex()`       | Set the `tabindex` attribute.                                                                   |
 | `title()`          | Set the `title` attribute.                                                                      |
@@ -163,8 +162,9 @@ The following methods are available for customizing the `HTML` output:
 Refer to the [Validate Tests](https://github.com/php-forge/html/blob/main/tests/Input/Time/ValidateTest.php) for
 comprehensive examples.
 
-| Method         | Description                                                                                         |
-| -------------- | --------------------------------------------------------------------------------------------------- |
-| `max()`        | Set the `max` attribute.                                                                            |
-| `min()`        | Set the `min` attribute.                                                                            |
-| `required()`   | Set the `required` attribute.                                                                       |
+| Method      | Description                                                                                            |
+| ----------- | ------------------------------------------------------------------------------------------------------ |
+| `max()`     | Set the `max` attribute.                                                                               |
+| `min()`     | Set the `min` attribute.                                                                               |
+| `step()`    | Set the `step` attribute.                                                                              |
+| `required()`| Set the `required` attribute.                                                                          |
