+++ tests/Input/Date/AttributeTest.php
@@ -175,16 +175,6 @@ final class AttributeTest extends TestCase
         );
     }
 
-    public function testStep(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <input id="date-6582f2d099e8b" type="date" step="1">
-            HTML,
-            Date::widget()->id('date-6582f2d099e8b')->step('1')->render()
-        );
-    }
-
     public function testStyle(): void
     {
         Assert::equalsWithoutLE(
