+++ tests/Input/Datetime/ValidateTest.php
@@ -33,6 +33,16 @@ final class ValidateTest extends TestCase
         );
     }
 
+    public function testStep(): void
+    {
+        Assert::equalsWithoutLE(
+            <<<HTML
+            <input id="datetime-6582f2d099e8b" type="datetime" step="1">
+            HTML,
+            Datetime::widget()->id('datetime-6582f2d099e8b')->step('1')->render()
+        );
+    }
+
     public function testRequired(): void
     {
         Assert::equalsWithoutLE(
