+++ tests/Input/DatetimeLocal/AttributeTest.php
@@ -173,16 +173,6 @@ final class AttributeTest extends TestCase
         );
     }
 
-    public function step(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <input id="datetime-local-6582f2d099e8b" type="datetime-local" step="value">
-            HTML,
-            DatetimeLocal::widget()->id('datetime-local-6582f2d099e8b')->step('value')->render()
-        );
-    }
-
     public function testStyle(): void
     {
         Assert::equalsWithoutLE(
