+++ tests/Input/Week/AttributeTest.php
@@ -173,16 +173,6 @@ final class AttributeTest extends TestCase
         );
     }
 
-    public function testStep(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <input id="week-6582f2d099e8b" type="week" step="1">
-            HTML,
-            Week::widget()->id('week-6582f2d099e8b')->step(1)->render()
-        );
-    }
-
     public function testStyle(): void
     {
         Assert::equalsWithoutLE(
